+++ src/Indice.Features.Messages.Core/Services/MessageEventQueue.cs
@@ -4,16 +4,16 @@ using Microsoft.Extensions.Options;
 
 namespace Indice.Features.Messages.Core.Services;
 /// <summary>Queue for storing campaign events.</summary>
-public class CampaignEventQueue
+public class MessageEventQueue
 {
-    private readonly CampaignStatisticOptions _statisticOptions;
+    private readonly AnalyticsOptions _Options;
     private readonly Channel<MessageEvent> _queue;
     private const int QUEUE_CHANNEL_CAPACITY = 100;
 
-    /// <summary>Initializes a new instance of the <see cref="CampaignEventQueue"/> class.</summary>
+    /// <summary>Initializes a new instance of the <see cref="MessageEventQueue"/> class.</summary>
     /// <param name="statisticOptions">Configuration for campaign statistics feature.</param>   
-    public CampaignEventQueue(IOptions<CampaignStatisticOptions> statisticOptions) {
-        _statisticOptions = statisticOptions.Value;
+    public MessageEventQueue(IOptions<AnalyticsOptions> statisticOptions) {
+        _Options = statisticOptions.Value;
         _queue = Channel.CreateBounded<MessageEvent>(new BoundedChannelOptions(QUEUE_CHANNEL_CAPACITY) {
             AllowSynchronousContinuations = false,
             SingleReader = true,
@@ -26,7 +26,7 @@ public class CampaignEventQueue
 
     /// <summary>Gets the writer for the queue.</summary>
     public ValueTask EnqueueAsync(MessageEvent lastActivityEntry) {
-        if (!_statisticOptions.EnableStatics) return ValueTask.CompletedTask;
+        if (!_Options.Enabled) return ValueTask.CompletedTask;
         return _queue.Writer.WriteAsync(lastActivityEntry);
     }
 }