+++ src/Indice.Features.Messages.Core/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.16.0] - 2025-09-19
+
+### Added New column
+```sql
+ALTER TABLE [msg].[Contact] 
+    ADD [Resolved]     bit      NULL
+GO
+CREATE NONCLUSTERED INDEX [IX_Contact_RecipientId_Resolved] 
+    ON [cmp].[Contact] ([RecipientId] ASC, [Resolved] ASC)
+GO
+CREATE NONCLUSTERED INDEX [IX_Campaign_CreatedAt]
+    ON [cmp].[Campaign] ([CreatedAt] ASC)
+GO
+```
 
 ## [8.1.10] - 2025-08-05